requirements.txt
PyYAML
scikit-learn
pandas
numpy
dvc








-e .